requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.6
python-multipart==0.0.20
pillow==11.1.0
numpy==2.2.2
torch==2.5.1
torchvision==0.20.1
mlflow==2.20.1
matplotlib==3.10.0
scikit-learn==1.6.1
google-cloud-storage==2.19.0
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.21.1
requests==2.32.3
dvc==3.59.1
pytest==8.3.4
httpx==0.28.1
flake8==7.1.1
pathspec==0.12.1
